32bit: change specific untyped constants and limits to allow compilation, reduce p2pool/sidechain memory usage during testing on 32-bit or short runs tests: disable specific usage of bench.ReportMetric due to TinyGo p2pool: Change AcceptTCP for Accept

+++ monero/address/wallet/interface_test.go
@@ -207,7 +207,7 @@ func testScanPayment[T curve25519.PointOperations](t *testing.T, wallet SpendWal
 				}
 
 				if scan.Amount != amount {
-					t.Fatalf("got amount %d, want %d", scan.Amount, amount)
+					t.Fatalf("got amount %d, want %d", scan.Amount, uint64(amount))
 				}
 			})
 		}
